+++ ironfish/src/wallet/index.ts
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 export * from './account'
-export * from './accounts'
+export * from './wallet'
 export { AccountsValue } from './database/accounts'
 export * from './validator'
 export * from './accountsdb'
